fix(config): update `prettier` configuration
Steps:
- Use the configuration of `nuxt-course` repository.
Change `proseWrap`, `quoteProps`, `singleQuote` and `trailingComma` properties to another values.
Add missing `singleAttributePerLine` property.
Use a matrix for override files.
Override `printWidth` property with another value for the `json` files.

Description:
- `"proseWrap": "preserve"` was changed to `"proseWrap": "always"`:
This means that automatic line wrapping will now be applied to paragraphs, preserving the original structure of the text.
- `"quoteProps": "consistent"` was changed to `"quoteProps": "preserve"`:
Now, the quotation style used in the properties will be preserved without applying automatic changes.
- `"singleQuote": true` was changed to `"singleQuote": false`:
This configuration was modified to avoid conflicts with the `eslint` configuration, which required double quotes instead of single quotes.
- The `"singleAttributePerLine": true` property was added:
This ensures that each attribute is written on a separate line, improving code readability and facilitating the identification of changes in individual attributes.
- `"trailingComma": "none"` was changed to `"trailingComma": "all"`:
Now, a comma will always be added at the end of the list of elements in arrays and objects, enhancing readability and facilitating the review of changes.
- Specific configuration was added for existing `.json` files using the `"printWidth": 0` property:
This means that there will be no line length limits for JSON files, preventing automatic adjustments.