feat(config): add configuration for use Dart sass/scss and compile files of `assets` folder
Steps:
- `vite` allows you to specify options that will be passed to `css` preprocessors, similar to how `vue-cli` handles `loaderOptions`.
To add additional data to all `scss` files in an `vite` project, you need to configure `css.preprocessorOptions` inside the `vite.config.js` file.
Note that the relative paths in `@import` inside `additionalData` must be relative to the location of the `vite.config.js` file or provide an alias that is configured in the `resolve.alias` option of the `vite` configuration file.
This configuration will automatically prepend the contents of `styles.scss` to each `scss` file in your project, allowing you to use the variables, mixins or functions defined in that file without needing to manually import them into each component.

Reason:
- The configuration used in the `vue.config.js` file has been added to the `vite.config.js` file, so that it does exactly the same as it did before with `@vue/cli`.