fix(index): add a `script` tag to link the `main.js` file inside the `index.html` file
Steps:
- The JavaScript application is no longer auto injected so we need to include it like this `<script type="module" src="/src/main.js"></script>` and remove the comment `<!-- built files will be auto injected -->`.

Reason:
- In projects configured with `@vue/cli` and `webpack`, the `JS` files, including the main application file (such as `main.js` or `app.js`), are automatically injected into the `index.html` during the build process. This is achieved through plugins such as `html-webpack-plugin`, which takes the bundle outputs generated by `webpack` and inserts them into the `index.html` in the appropriate places. This automatic injection facilitates dependency management and ensures that the scripts are loaded correctly, taking into account file hashing for cache control and the correct order of the scripts.
However, when you migrate to `vite`, this automatic script injection functionality is not handled in the same way. Being a more modern build tool and focused on leveraging standard `ESModules` features in browsers, it does not automatically inject scripts into the `index.html`. Therefore, to ensure that your application loads correctly, you need to explicitly specify the entry point of your application (usually `src/main.js` for `vue` applications) in the `index.html` using a `<script>` tag with the `type="module"` attribute.
It uses the native capability of modern browsers to work with `ESModules`, allowing for simpler, web standards-aligned development and deployment.
By explicitly stating the script in `index.html`, you have full control over the entry point of your application and how it loads.
Take advantage of the `ESModules` optimized development server, which allows for fast reloads and efficient `Hot Module Replacement (HMR)`.
By not relying on automatic script injection, the need for additional configurations and plugins is reduced, simplifying the process of building and deploying your application.