feat(config): update scripts `serve`, `build` and `lint` to use `vite` instead `vue-cli-service`
Steps:
- In the `package.json` file, remove the `serve` and `build` npm scripts from `vue-cli-service`.
```
"serve": "vue-cli-service serve",
"build": "vue-cli-service build",
```
- In the the `package.json` file, add the new `dev`, `build` and `preview` npm scripts from `vite`:
```
"dev": "vite",
"build": "vite build",
"preview": "vite preview"
```
- In the `package.json` file, the npm `lint` script has to use the `eslint` command directly instead of `vue-cli-service`.
`"eslint:fix": "vue-cli-service lint..."`

Reason:
- When migrating from `@vue/cli` to `vite`, replace the old `vue-cli-service` based commands with their `vite` equivalents in your `package.json` file:
The `dev` command in `vite`, is the equivalent of the `serve` command in `@vue/cli` projects. But with the difference that `vite` provides a much faster startup and a more efficient `Hot Module Replacement (HMR)` thanks to its modern and optimized approach based on native `ESModules`, which allows code updates to be reflected almost instantly in the browser.
The `build` command in `vite`, is the same in `@vue/cli` projects. It compiles the project for production, but under the hood, `vite` uses `Rollup` for the packaging process, resulting in an optimized package ready to be deployed in a production environment, providing a build process that is finely tuned for efficiency and optimizing the end result.
The `preview` command of `vite` does not have a direct equivalent in `@vue/cli`, but it is a very useful additional feature that `vite` offers to facilitate local production testing. It is used once you have built your project with the `build` command, to locally serve the built production files, by starting a local static server that serves the generated files in the `dist` folder. This is useful for testing the production version of your application in a local environment before deploying it.
Note that the command to spin up the development server is no longer `serve`,`vite` uses `dev` instead and `serve` is used to preview the production build locally.