fix(import): `stylelint` errors about the `scss/at-import-partial-extension-disallowed-list` rule
Errors:
- src/App.vue
 51:20  ✖  Unexpected extension ".scss"    scss/at-import-partial-extension-disallowed-list in imported partial name
 52:20  ✖  Unexpected extension ".scss"    scss/at-import-partial-extension-disallowed-list in imported partial name
 56:20  ✖  Unexpected extension ".scss"    scss/at-import-partial-extension-disallowed-list in imported partial name
3 problems (3 errors, 0 warnings)

Solution:
- Remove the unnecessary file extension `.scss`.

Reason:
- In Stylelint, the `scss/at-import-partial-extension-disallowed-list` rule is used to prevent the use of file extensions in SASS/SCSS `@import`.
The reason for this rule is to promote consistency and to comply with the convention in SASS that file extensions are not required when importing partials.
SASS allows files to be imported without specifying the `.scss` extension, and doing so is considered good practice since SASS knows to look for files with that extension by default.
Also, omitting the extension allows some flexibility in case the file extension changes to `.sass` or if indexing with `_index.scss` is being used.

References:
- https://github.com/stylelint-scss/stylelint-scss/tree/master/src/rules/at-import-partial-extension-disallowed-list