fix(import): remove magic comments like `webpackChunkName`
Steps to follow:
- In this case, there are magic comments like `webpackChunkName` in the `router/index.js` file, because we put them back in the `commit` above.
At some point in the development of the application, they were removed so we put them back to remove them again in this change to explain why they are not needed and should be removed when we use `vite` and not `@vue/cli`.

Reason:
- Remove all magic comments (`/* webpackChunkName: "about" */`) to name your dynamic imports as they are `webpack` specific comments and mean nothing to `vite`.
Instead, `vite` will automatically name your chunk based on the original `.vue` file name combined with a cache hash, like this: `About.37a9fa9f.js`.