feat(env): change `process.env.BASE_URL` to `import.meta.env.BASE_URL`
Steps:
- Replace in the `router/index.js` file the `process.env.BASE_URL` instance with `import.meta.env.BASE_URL`, ensuring compatibility with `vite` environment variable handling.

Reason:
- The `.env` file is not used in this project at this time.
But it is important to know that the `.env` naming convention is the same for `vite` (`.env`, `.env.local`, `.env.[mode]`, `.env.[mode].local`).
- However, you can no longer access them from `vue` with a `process.env.` variable (`process.env.BASE_URL`), instead you must use `import.meta.env.` (`import.meta.env.BASE_URL`).
Vite uses `import.meta.env` to inject environment variables, as opposed to `process.env` which is used in tools such as `webpack`.
This change ensures that access to environment variables follows `vite` conventions and maintains consistency with your `ESModules` system, allowing variables to be used in both development and production.

References:
- https://vueschool.io/articles/vuejs-tutorials/how-to-migrate-from-vue-cli-to-vite/