fix(env): `eslint` warnings when using `process.`
Error:
- `process` is not defined. eslint (no-undef).
Warnings found in `eslint.cjs` file in `no-console` and `no-debugger` rules.
Warning found in `gh-pages-deploy.js` file when using it in `process.exit(1)`.

Solution:
- Register `process` as a `readonly` global variable in the `eslint` configuration, to avoid warnings about its use in environment checks.
To fix the `warning` that appears when `process.env.NODE_ENV` is used in the `eslint` configuration, you need to add a `global` statement to the beginning of the `.eslintrc.cjs` file.
This will tell `eslint` to recognize `process.env` as a global variable that should not be modified and prevent it from marking it as `undefined`, allowing it to be used throughout the configuration without `eslint` displaying warnings.
This change resolves warnings in both the `.eslintrc.cjs` file and the `gh-pages-deploy.js` file, where `process` is used.

Reason:
- In your `.eslintrc.cjs` file, you should still use `process.env.NODE_ENV` since `import.meta.env` is not available outside the context of modules and files handled by `vite`.
The `eslint` configuration generally runs in a `node` environment where `process.env.NODE_ENV` is the standard way to access environment variables.
- This change resolves `eslint` warnings by explicitly recognizing `process` as part of the global scope within `eslint`. It ensures that `NODE_ENV` checks for environment-specific rules (such as `no-console` and `no-debugger`) do not result in undefined references, facilitating development and production optimizations.