feat(config): change the `server.port` from `5173` to `3000`
Steps:
- Define the `server` object within `vite.config.js`.
- Set the `port` option to `3000`, replacing the default port `5173`.

Reason:
- In `vite` by default, when running the `npm run dev` command, the `server.host` is `localhost` and the `server.port` is `5173`.
But we want to use the same local port we had with `@vue/cli`, since anyone who has worked on this project before expects port `3000` to be used.
So we change the configuration in the `vite.config.js` file, so that the path to display the application locally changes from `http://localhost:5173/` to `http://localhost:3000/`.

References:
- https://vitejs.dev/config/server-options#server-host
- https://vitejs.dev/config/server-options#server-port