fix(App): the `home` page has broken styles
Error:
- main.js:19 [Vue warn]: Error in callback for immediate watcher "$route": "TypeError: Cannot read properties of null (reading 'toLowerCase')"
found in
---> <App> at /Users/beatrizsmerino/Sites/localhost/vue-editor/src/App.vue
- main.js:19 TypeError: Cannot read properties of null (reading 'toLowerCase')
at VueComponent.handler (App.vue:28:1)
- main.js:15 [Vue warn]: Error in callback for watcher "$route": "TypeError: Cannot read properties of null (reading 'toLowerCase')"
found in
---> <App> at /Users/beatrizsmerino/Sites/localhost/vue-editor/src/App.vue
- main.js:15 TypeError: Cannot read properties of null (reading 'toLowerCase')
at VueComponent.handler (App.vue:26:1)

Solution:
- Ensure route names exist before string manipulation, to avoid null reference errors.
Added a check for the existence of the `from.name` and `to.name` before calling the `toLowerCase()` method, ensuring we only try to manipulate classes with valid strings.

Reason:
- The application was encountering `null` reference errors when the route's `name` property was undefined. This occurred during route transitions where the previous or next route did not have a defined `name` property, leading to JavaScript errors when attempting to call `toLowerCase()` of `null`.
The change ensures that the dynamic class updates on the HTML element are applied only on routes with valid names, preserving the integrity of the application's runtime behavior and preventing console errors.