fix(EditorJSON): cannot access to the `editors` page
Error:
- vue-router.js?v=9325602b:1692 SyntaxError: The requested module '/node_modules/vue-json-editor/assets/jsoneditor.js?v=9325602b' does not provide an export named 'default' (at vue-json-editor.vue:3:8)
- EditorJSON.vue:89 [Vue warn]: Failed to resolve async component: () => import("/node_modules/vue-json-editor/vue-json-editor.vue")
Reason: SyntaxError: The requested module '/node_modules/vue-json-editor/assets/jsoneditor.js?v=9325602b' does not provide an export named 'default'

Solution:
- Dynamically imports the `VueJsonEditor` component.

Reason:
- With this change we can now navigate to the `editors` page although the problem about the `VueJsonEditor` component persists.
It seems that the `vue-json-editor` package is not compatible with `vite` because it is not prepared for use with `ESModules`.
This is a temporary solution to be able to access the page, but you will have to replace the `vue-json-editor` package with another one that has the same functionality and is compatible.

References:
- https://www.npmjs.com/package/vue-json-editor
- https://github.com/dirkliu/vue-json-editor/issues/42
- https://github.com/dirkliu/vue-json-editor/issues/51