feat(config): remove config for `svg-symbol-sprite-loader` and add for `vite-plugin-svg-icons`
Steps:
- In the `src/assets/images/icons/icons-import.js` file remove the `svg-symbol-sprite-loader` package import and add the `vite-plugin-svg-icons` package import instead.
This is in line with Vite's best practices, which recommend using the `virtual:` prefix for plugin-provided virtual modules.
The `icons-import.js` file is imported into the `main.js` file, so it is as if the `vite-plugin-svg-icons` package is imported directly into the `main.js` file.
- In the `vite.config.js` file, add the import of `createSvgIconsPlugin` function from the `vite-plugin-svg-icons` package.
It uses the `createSvgIconsPlugin` function inside the `plugins` property to generate the svg sprite map as soon as the `npm run dev` command is executed.
- To the `createSvgIconsPlugin` function, add an object with the `iconDirs` and `symbolID` options to customize it.
The `iconDirs` option will have as value an array with the path to the folder containing the svg icons that will be automatically added to the sprite.
The `symbolId` option will have as value a string with the name of the icon to be able to call it from the sprite and display it.