build(deps): remove the configuration of `vite-plugin-eslint` and `vite-plugin-stylelint` packages
Steps:
- In the `vite.config.js` file:
Remove the imports about `vite-plugin-eslint` and `vite-plugin-stylelint` packages.
Remove the execution of these packages (`eslint()` and `stylelint()`) from the `plugins` property.

Reason:
- `Husky` efficiently handles formatting and error correction before uploading changes to the repository, eliminating the need for `vite-plugin-eslint` and `vite-plugin-stylelint` packages.
This optimizes the performance of `vite`, since such plugins are not guaranteed to work correctly on every file save. Simplifying this part of the workflow improves efficiency and keeps the code base clean and consistent.