fix(JsonEditorVue): don't use `this` to access `methods`, `props` or `data` in `Composition API`
Errors:
- [Vue warn]: Error in mounted hook: "TypeError: Cannot read properties of null (reading 'getData')"
- TypeError: Cannot read properties of null (reading 'getData')
- [Vue warn]: Error in mounted hook: "TypeError: Cannot read properties of null (reading 'apiUrl')"
- TypeError: Cannot read properties of null (reading 'apiUrl')
- Uncaught (in promise) TypeError: Cannot set properties of null (setting 'json')

Solution:
- Do not use `this` to access `props`, methods or reactive variables in the `Composition API`.
Browser console errors indicate `getData`, `apiUrl` and `json` cannot be read. This is because in the `Composition API`, you do not have access to `this` inside `setup()`. This is why all reactive properties and methods must be declared inside `setup()` and returned at the end.

Steps:
- Besides defining and returning functions as constants inside `setup()`, to call them you don't need `this.`, you can call them directly:
That is, to execute the `getData` method use `getData()` instead of `this.getData()`.
- To use the `props`, you need to pass them as a parameter to the `setup()` function and to access them directly using `props.`:
That is, to access the value of the `apiUrl` prop use `props.apiUrl` instead of `this.apiUrl`.
- In addition to defining and returning variables as constants inside `setup()` using `ref()`, to get the value of reactive references use `.value`.
That is, to get the value of the data `json`, since `json` has been defined as `ref()` use `json.value` instead of `this.json`.