fix(scss): wrong naming conventions for comments and titles
Error:
- Each scss file must contain a title at the top that complies with the following:
All title text must be in uppercase.
The first line of the title must contain the name of the folders in which the file is located separated by hyphens.
The second line of the title must contain the name of the file.

Solution:
- In the `_base-reset.scss` file:
Change `GENERICS` to `BASE`.
Change `reset-personal` to `RESET`.
- In the `_base-global.scss` file:
Change `COMPONENTS` to `BASE`.
Change `GENERICS` to `GLOBAL`.
- In the `_abstracts-variables-fonts.scss` file:
Change `SETTINGS - VARIABLES` to `ABSTRACTS - VARIABLES`.
- In the `_abstracts-variables-colors.scss` file:
Change `ABSTRACTS VARIABLES` to `ABSTRACTS - VARIABLES`.
- In the `_abstracts-variables-breakpoints.scss` file:
Change `SETTINGS` to `ABSTRACTS - VARIABLES`.
- In the `_abstracts-mixins-media.scss` file:
Change `MIXIN` to `ABSTRACTS - MIXINS`.
Change `MEDIA QUERIES` to `MEDIA`.