fix(theme): when emulating `prefer-colour-scheme: light` some things are not displayed
Error:
- In the light theme, the `page__title`, `page-nav__link` and `button` selectors are not displayed.
Their `colour` is white like the background of the screen, so they cannot be distinguished.

Solution:
- Use css variables so that the media query `prefers-color-scheme: dark` is applied.
Create the css variables `--color-link` and `--color-button` to switch the colour between the light and dark theme.
Instead of using `$color-white` use `--color-heading` for `page__title`, `--color-link` for `page-nav__link` and `--color-button` for `button`.

References:
- https://stackoverflow.com/a/59223868/10855837