fix(deploy): missing `gh-pages-deploy.mjs` file when running the `deploy` command
Error:
- Cannot find module '/home/runner/work/vue-editor/vue-editor/gh-pages-deploy.mjs'.

Reason:
- This occurs because the `deploy` command attempts to execute the `gh-pages-deploy.mjs` file, which no longer exists.
At some point, the `gh-pages-deploy.mjs` file was renamed to `gh-pages-deploy.js`, but the command was not updated to reflect this change.
- This new version of the project uses `vite`, which takes advantage of modern JavaScript features, such as native `ES modules`.
Therefore, it is no longer necessary to use the `.mjs` extension for `js` files, as the `package.json` file configuration has been changed to `ES Modules` by default.
When it is necessary to use `CommonJS`, the `js` files will be renamed with the `.cjs` extension.

Solution:
- Update the `deploy` command to reference the new file name `gh-pages-deploy.js`.