fix(deps): error about `vue/require-explicit-emits` rule when updating `eslint-plugin-vue` to 10.0.0
Error:
- vue-editor/src/components/UI/UIButton.vue
  34:16  warning  The "click" event has been triggered but not declared on `emits` option  `vue/require-explicit-emits`

Reason:
- The error comes from the upgrade of `eslint-plugin-vue` to version `10.0.0`, which now applies the `vue/require-explicit-emits` rule.
This rule requires that all events emitted with `$emit` be explicitly declared in the `emits` option of the component.

Solution:
- Add the `emits` option to the configuration of your `UIButton.vue` component.
The `"emits": ["click"]` key indicates that the `UIButton` component intends to emit an event called `“click”`, complying with the `vue/require-explicit-explicit-emits` rule.
This avoids the ESLint warning and ensures that the code is more declarative and easier to maintain.